[Tests] Test refactoring

+++ src/hir/hirgen/body/block_hirgen.c
@@ -93,7 +93,7 @@ static int _navigation_handler(ast_node_t* node, hir_ctx_t* ctx, sym_table_t* sm
         case FUNC_TOKEN:       return HIR_generate_function_block(node, ctx, smt);
         case EXIT_TOKEN:       return HIR_generate_exit_block(node, ctx, smt);
         case CALLING_TOKEN:
-        case CALL_TOKEN:       return (int)HIR_generate_funccall(node, ctx, smt, 0);
+        case CALL_TOKEN:       return (int)((long)HIR_generate_funccall(node, ctx, smt, 0));
         case LOOP_TOKEN:       return HIR_generate_loop_block(node, ctx, smt);
         case BREAK_TOKEN:      return HIR_generate_break_block(ctx);
         case WHILE_TOKEN:      return HIR_generate_while_block(node, ctx, smt);
@@ -102,7 +102,7 @@ static int _navigation_handler(ast_node_t* node, hir_ctx_t* ctx, sym_table_t* sm
         case RETURN_TOKEN:     return HIR_generate_return_block(node, ctx, smt);
         case EXTERN_TOKEN:     return HIR_generate_extern_block(node, ctx);
         case ASSIGN_TOKEN:     return HIR_generate_assignment_block(node, ctx, smt);
-        case SYSCALL_TOKEN:    return (int)HIR_generate_syscall(node, ctx, smt, 0);
+        case SYSCALL_TOKEN:    return (int)((long)HIR_generate_syscall(node, ctx, smt, 0));
         case BREAKPOINT_TOKEN: return HIR_generate_breakpoint_block(node, ctx);
         default: break;
     }
